nginx upgrade fix
nginx upgrade fix

+++ nginx_upgrade.sh
@@ -35,9 +35,15 @@ sudo cp -r /tmp/security.conf /etc/nginx/cryptopool.builders
 sudo cp -r /tmp/letsencrypt.conf /etc/nginx/cryptopool.builders
 
 # Removing default nginx site configs.
+if [ -f /etc/nginx/conf.d/default.conf ]; then
 sudo rm -r /etc/nginx/conf.d/default.conf
+fi
+if [ -f /etc/nginx/sites-available/default ]; then
 sudo rm -r /etc/nginx/sites-available/default
+fi
+if [ -f /etc/nginx/sites-enabled/default ]; then
 sudo rm -r /etc/nginx/sites-enabled/default
+fi
 
 restart_service nginx;
 wait $!
